+++ bootstrap/app.php
@@ -4,6 +4,7 @@ use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Configuration\Exceptions;
 use Illuminate\Foundation\Configuration\Middleware;
 use App\Http\Middleware\CheckRole;
+use App\Http\Middleware\LogRootActivity;
 
 return Application::configure(basePath: dirname(__DIR__))
     ->withRouting(
@@ -23,7 +24,8 @@ return Application::configure(basePath: dirname(__DIR__))
            
         ]);
          $middleware->alias([
-                'role' => CheckRole::class,
+                'role'     => CheckRole::class,
+                'log-root' => LogRootActivity::class,
          ]);
     })
     ->withExceptions(function (Exceptions $exceptions): void {